pyproject.toml
[tool.poetry]
name = "stock-backtesting-tool"
version = "0.1.0"
description = ""
authors = ["Connor Young <connoryoungkicker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
pydantic = "^2.6.1"
uvicorn = "^0.27.1"
requests = "^2.31.0"
yfinance = "^0.2.36"
backtesting = "^0.3.3"
pandas-datareader = "^0.10.0"
pendulum = "^3.0.0"
requests-cache = "^1.2.0"
exchange-calendars = "^4.5.3"

[tool.poetry.scripts]
my-script = "my_module:main"
get-flagged-stocks = "my_module:HelloWorld"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
